Add figureIIIF filter to extract IIIF properties from figure data and use in canvasPanel and choices shortcodes
