+++ packages/11ty/_includes/components/figure/index.js
@@ -1,4 +1,5 @@
 const caption = require('./caption')
+const choices = require('./choices')
 const image = require('./image')
 const imageservice = require('./imageservice')
 const label = require('./label')
@@ -11,6 +12,7 @@ const youtube = require('./youtube')
 
 module.exports = {
   figurecaption: caption,
+  figurechoices: choices,
   figureimage: image,
   figurelabel: label,
   figuremodallink: modallink,
