+++ packages/11ty/_plugins/filters/index.js
@@ -1,5 +1,6 @@
 const capitalize = require('./capitalize')
 const isImageService = require('./isImageService')
+const figureIIIF = require('./figureIIIF')
 const fullname = require('./fullname')
 const getContributor = require('./getContributor')
 const getFigure = require('./getFigure')
@@ -12,6 +13,8 @@ module.exports = function(eleventyConfig, options) {
   eleventyConfig.addFilter('capitalize', (string) => capitalize(string))
 
   eleventyConfig.addFilter('isImageService', (figure, options) => isImageService(figure, options))
+
+  eleventyConfig.addFilter('figureIIIF', (figure, options) => figureIIIF(eleventyConfig, figure, options))
 
   eleventyConfig.addFilter('fullname', (person, options) => fullname(person, options))
